+++ HISTORY.md
@@ -6,7 +6,9 @@
 
 ## 1.2.0
 
-
+* [#13: Add recursion detection when injecting constructor parameters](https://github.com/RotenInformatik/AbstractionsDotNet/issues/13)
+* [#14: Add NullLogger and make ILogger registration optional](https://github.com/RotenInformatik/AbstractionsDotNet/issues/14)
+* [#15: SimpleContainer does not properly adhere Transient registrations](https://github.com/RotenInformatik/AbstractionsDotNet/issues/15)
 
 ## 1.1.0